% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataCoord}
\alias{dataCoord}
\title{Data Coord}
\format{
Une data frame avec quatres variables :
\describe{
\item{\code{Year}}{Année}
#' \item{\code{Circuit}}{Nom des circuits où sont les Grands Prix}
\item{\code{Lat}}{Latitude}
\item{\code{Lng}}{Longtitude}
}
}
\usage{
dataCoord
}
\description{
Données sur chaque Grand Prix de 1950 à 2020
}
\keyword{datasets}
