% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDrivers}
\alias{dataDrivers}
\title{Data Drivers}
\format{
Une data frame avec cinq variables:
\describe{
\item{\code{Year}}{Années}
\item{\code{Postion}}{Position finale des pilotes par année}
\item{\code{Driver}}{Nom des pilotes}
\item{\code{Car}}{Nom des écuries}
\item{\code{Points}}{Nombre de points gagnés par les pilotes par année}
}
}
\usage{
dataDrivers
}
\description{
Classement final des pilotes de 1950 à 2020
}
\keyword{datasets}
