% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataDriversParRaces}
\alias{dataDriversParRaces}
\title{Data Drivers par Races}
\format{
Une data frame avec sept variables :
\describe{
\item{\code{Indice}}{Indice des lignes}
\item{\code{Driver}}{Nom des pilotes}
\item{\code{GrandPrix}}{Nom des Grand Prix}
\item{\code{Year}}{Années}
\item{\code{Car}}{Nom des écuries}
\item{\code{Position}}{Positions finales des pilotes}
\item{\code{Points}}{Points gagnés par les pilotes}
}
}
\usage{
dataDriversParRaces
}
\description{
Donne le résultat de toutes les courses de F1 de 1950 à 2020
}
\keyword{datasets}
