% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataFL}
\alias{dataFL}
\title{Data Fastest Lap (DataFL)}
\format{
Une data frame avec cinq variables :
\describe{
\item{\code{Year}}{Années}
\item{\code{GrandPrix}}{Nom de Grand Prix}
\item{\code{Driver}}{Nom des pilotes}
\item{\code{Car}}{Nom des écuries}
\item{\code{Time}}{Temps du tour le plus rapide}
}
}
\usage{
dataFL
}
\description{
Tour le plus rapide pour chaque Grand Prix de 1950-2020
}
\keyword{datasets}
