% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataRaces}
\alias{dataRaces}
\title{Data Races}
\format{
Une data frame avec sept variables :
\describe{
\item{\code{GrandPrix}}{Nom des Grand Prix}
\item{\code{Date}}{Date des Grand Prix}
\item{\code{Year}}{Année}
\item{\code{Winner}}{Vainqueur du Grand Prix}
\item{\code{Car}}{Nom des écuries}
\item{\code{Laps}}{Nombre de tours}
\item{\code{Time}}{Temps du vainqueur}
}
}
\usage{
dataRaces
}
\description{
Données sur chaque Grand Prix de 1950 à 2020
}
\keyword{datasets}
