% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataStartingGrid}
\alias{dataStartingGrid}
\title{Data Starting Grid}
\format{
Une data frame avec cinq variables :
\describe{
\item{\code{Year}}{Années}
\item{\code{GrandPrix}}{Nom de Grand Prix}
\item{\code{Position}}{Positions de départ}
\item{\code{Driver}}{Nom des pilotes}
\item{\code{Car}}{Nom des écuries}
}
}
\usage{
dataStartingGrid
}
\description{
Position de départ de chaque pilotes pour chaque Grand Prix de 1950 à 2020
}
\keyword{datasets}
