% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dataTeams}
\alias{dataTeams}
\title{Data Teams}
\format{
Une data frame avec quatre variables:
\describe{
\item{\code{Year}}{Années}
\item{\code{Position}}{Position finale des écuries par année}
\item{\code{Team}}{Nom des écuries}
\item{\code{Points}}{Nombres de points gagnés par les écuries par année}
}
}
\usage{
dataTeams
}
\description{
Classement des écuries de 1958 à 2020
}
\keyword{datasets}
