% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNF_S_freq.R
\name{DNF_S_freq}
\alias{DNF_S_freq}
\title{Nombre de DNF et DNS de chaque pilote}
\usage{
DNF_S_freq(dataDrivers)
}
\arguments{
\item{dataDrivers}{data frame}
}
\value{
graphique
}
\description{
Nombre de DNF et DNS de chaque pilote
}
