% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNFDriver.R
\name{DNFDriver}
\alias{DNFDriver}
\title{Nombre de DNF et DNS de chaque pilote}
\usage{
DNFDriver(data_driver_races)
}
\arguments{
\item{data_driver_races}{data}
}
\value{
graphique
}
\description{
Nombre de DNF et DNS de chaque pilote
}
