% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNFGrandPrix.R
\name{DNFGrandPrix}
\alias{DNFGrandPrix}
\title{Nombre de DNF et DNS de chaque grand prix}
\usage{
DNFGrandPrix(data_race_driver)
}
\arguments{
\item{data_race_driver}{data frame}
}
\value{
graphique
}
\description{
Affichage d’un diagramme en barre, réalisé sous plotly et ggplot2, représentant le nombre d'abandons par grand prix sur une saison. Les données doivent être issues de la data frame dataDriversParRaces.
}
