% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNFTeam.R
\name{DNFTeam}
\alias{DNFTeam}
\title{Pourcentage de victoires par écurrie}
\usage{
DNFTeam(data)
}
\arguments{
\item{data}{data frame}
}
\value{
graphique
}
\description{
Affichage d’un diagramme circulaire, réalisé sous plotly, représentant le pourcentage d'abandon pour chaque écuries lors d'une saison. Les données doivent être issues de la data frame dataDriversParRaces.
}
