% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointDriverGP.R
\name{PointDriverGP}
\alias{PointDriverGP}
\title{Représentation des points gagnés par pilote par grand prix}
\usage{
PointDriverGP(data_driver_race, data_race)
}
\arguments{
\item{data_driver_race}{data frame}

\item{data_race}{dataframe}
}
\value{
graphique
}
\description{
Affichage d’un graphique réalisé avec plotly et ggplot2 représentant le nombre de points gagnés de chaque pilote par grand prix sur une saison. Les données doivent être issues des data frames dataDriversParRaces et dataRaces.
}
