% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointEcurieGP.R
\name{PointEcurieGP}
\alias{PointEcurieGP}
\title{Représentation de la somme des points gagnés par écurie par grand prix}
\usage{
PointEcurieGP(data_race_driver)
}
\arguments{
\item{data_race_driver}{data frame}
}
\value{
graphique
}
\description{
Affichage d’un graphique avec plotly et ggplot2 représentant les points gagnés de chaque écurie par grand prix sur une saison. Les données doivent être issues de la data frame dataDriversParRaces.
}
