% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosFinalePilote.R
\name{PosFinalePilote}
\alias{PosFinalePilote}
\title{Représentation de la position finale des pilotes ayant fait le tour le plus rapide}
\usage{
PosFinalePilote(data_driver_race, data_fl, data_races, data_driver)
}
\arguments{
\item{data_driver_race}{data frame}

\item{data_fl}{data frame}

\item{data_races}{data frame}

\item{data_driver}{data frame}
}
\value{
diagramme en barre
}
\description{
Affichage d’un diagramme en barre, réalisé sous plotly et ggplot2, représentant la fréquence de la position d'arrivée du pilote ayant fait le tour le plus rapide. Les données doivent être issues des data frames: dataDriversParRaces, dataFL, dataRaces, dataDrivers.
}
