% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pts_moyens_Driver_GP.R
\name{Pts_moyens_Driver_GP}
\alias{Pts_moyens_Driver_GP}
\title{Nombre de points moyens par grand prix de chaque pilote}
\usage{
Pts_moyens_Driver_GP(datadrivers, dataraces)
}
\arguments{
\item{datadrivers}{data frame}

\item{dataraces}{data frame}
}
\value{
graphique
}
\description{
Nombre de points moyens par grand prix de chaque pilote
}
