% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTeamAnalyse.R
\name{DataTeamAnalyse}
\alias{DataTeamAnalyse}
\title{Représentation des données pour les écuries}
\usage{
DataTeamAnalyse(data)
}
\arguments{
\item{data}{data frame}
}
\value{
donut chart
}
\description{
#' La data frame en paramètre doit faire parti des données représentées
sur la page https://www.formula1.com/en/results.html/2021/races.html
}
