% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DataTeamAnalyse.R
\name{DataTeamAnalyse}
\alias{DataTeamAnalyse}
\title{Représentation des données pour les écuries}
\usage{
DataTeamAnalyse(data)
}
\arguments{
\item{data}{data frame}
}
\value{
donut chart
}
\description{
Affichage d’un graphique sous plotly représentant le pourcentage de victoire des écuries sur une saison. Les données doivent être issues de la data frame dataTeams.
}
