% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNFDriver.R
\name{DNFDriver}
\alias{DNFDriver}
\title{Nombre de DNF et DNS de chaque pilote}
\usage{
DNFDriver(data_driver_races)
}
\arguments{
\item{data_driver_races}{data}
}
\value{
graphique
}
\description{
Affichage d’un diagramme en barre réalisé sous plotly et ggplot2, représentant le nombre d'abandons par pilotes sur une saison. Les données doivent être issues de la data frame dataDriversParRaces.
}
