% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNFGrandPrix.R
\name{DNFGrandPrix}
\alias{DNFGrandPrix}
\title{Nombre de DNF et DNS de chaque grand prix}
\usage{
DNFGrandPrix(data_race_driver)
}
\arguments{
\item{data_race_driver}{data frame}
}
\value{
graphique
}
\description{
Nombre de DNF et DNS de chaque grand prix
}
