% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointDriverGP.R
\name{PointDriverGP}
\alias{PointDriverGP}
\title{Représentation des points gagnes par pilote par grand prix}
\usage{
PointDriverGP(data_driver_race, data_race)
}
\arguments{
\item{data_driver_race}{data frame}

\item{data_race}{dataframe}
}
\value{
graphique
}
\description{
Représentation des points gagnes par pilote par grand prix
}
