% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PointEcurieGP.R
\name{PointEcurieGP}
\alias{PointEcurieGP}
\title{Représentation de la somme des points gagnés par écurie par grand prix}
\usage{
PointEcurieGP(data_race_driver)
}
\arguments{
\item{data_race_driver}{data frame}
}
\value{
graphique
}
\description{
La data frame en paramètre doit faire parti des données représentées
sur la page https://www.formula1.com/en/results.html/2021/races.html
}
