% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosDepArr.R
\name{PosDepArr}
\alias{PosDepArr}
\title{Comparaison des positions de depart et d'arrivee des pilotes}
\usage{
PosDepArr(data_sg, data_driver_race, grand_prix)
}
\arguments{
\item{data_sg}{data frame}

\item{data_driver_race}{data frame}

\item{grand_prix}{data frame}
}
\value{
graphique
}
\description{
Affichage d’un nuage de points, réalisé sous plotly et ggplot2, représentant la position d'arrivée d'un pilote en fonction de sa position de départ sur un grand prix. Les données doivent être issues des data frames dataStartingGrid et dataDriversParRaces.
}
