% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PosFinalePilote.R
\name{PosFinalePilote}
\alias{PosFinalePilote}
\title{Représentation de la position finale des pilotes ayant fait le tour le plus rapide}
\usage{
PosFinalePilote(data_driver_race, data_fl, data_races, data_driver)
}
\arguments{
\item{data_driver_race}{data frame}

\item{data_fl}{data frame}

\item{data_races}{data frame}

\item{data_driver}{data frame}
}
\value{
diagramme en barre
}
\description{
Les datas frames en paramètre doit faire parti des données représentées
sur la page https://www.formula1.com/en/results.html/2021/races.html
}
