% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pts_moyens_Driver_GP.R
\name{Pts_moyens_Driver_GP}
\alias{Pts_moyens_Driver_GP}
\title{Nombre de points moyens de chaque pilote sur une saison}
\usage{
Pts_moyens_Driver_GP(datadrivers, dataraces)
}
\arguments{
\item{datadrivers}{data frame}

\item{dataraces}{data frame}
}
\value{
graphique
}
\description{
Affichage d’un diagramme en barre, réalisé sous plotly et ggplot2, représentant le nombre de points moyens de chaque pilote sur une saison. Les données doivent être issues des data frames: dataRaces et dataDrivers.
}
